+++ Runtime/CommonComponent/PhaseDefinition.cs
@@ -2,9 +2,11 @@ using UnityEngine;
 
 namespace net.rs64.TexTransTool
 {
-    [AddComponentMenu("TexTransTool/Group/TTT PhaseDefinition")]
+    [AddComponentMenu(TexTransBehavior.TTTName + "/" + PDMenuPath)]
     public sealed class PhaseDefinition : TexTransGroup
     {
+        internal const string PDName = "TTT PhaseDefinition";
+        internal const string PDMenuPath = TexTransGroup.FoldoutName + "/" + PDName;
         public TexTransPhase TexTransPhase;
     }
 }